Changes argocd config lookup to external data source (#10)
* Changes argocd config lookup to external data source

+++ scripts/delete-tls-secret.sh
@@ -3,6 +3,13 @@
 NAMESPACE="$1"
 SECRET_NAME="$2"
 
+export PATH="${BIN_DIR}:${PATH}"
+
+if ! command -v kubectl 1> /dev/null 2> /dev/null; then
+  echo "kubectl cli not found"
+  exit 1
+fi
+
 if kubectl get secret -n "${NAMESPACE}" "${SECRET_NAME}"; then
   kubectl delete secret -n "${NAMESPACE}" "${SECRET_NAME}"
 else
